… etc/grafana/dashboards/foyer.json