@MrCroxx 2024-05-19
## Benchmark

### Command

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-bench --dir /p44pro/foyer --mem 64 --disk 1024 --file-size 64 --get-range 10000 --flushers 4 --reclaimers 4 --time 300 --writers 64 --w-rate 8 --admission-rate-limit 500 --readers 1024 --r-rate 1 --runtime --compression none --distribution zipf --distribution-zipf-s 0.8 --metrics
```

### main

```
Total:
disk total iops: 11137.3
disk total throughput: 599.3 MiB/s
disk read iops: 6969.3
disk read throughput: 138.5 MiB/s
disk write iops: 4168.0
disk write throughput: 460.9 MiB/s
insert iops: 6898.6/s
insert throughput: 431.2 MiB/s
insert lat p50: 3us
insert lat p90: 5us
insert lat p99: 8us
insert lat p999: 26us
insert lat p9999: 154us
insert lat p99999: 235us
insert lat pmax: 409us
get iops: 17199.6/s
get miss: 5.38%
get throughput: 1.0 GiB/s
get hit lat p50: 3us
get hit lat p90: 259us
get hit lat p99: 491us
get hit lat p999: 3263us
get hit lat p9999: 6431us
get hit lat p99999: 6911us
get hit lat pmax: 7455us
get miss lat p50: 158us
get miss lat p90: 423us
get miss lat p99: 735us
get miss lat p999: 1119us
get miss lat p9999: 1303us
get miss lat p99999: 1367us
get miss lat pmax: 1431us
```

### PR

```
Total:
disk total iops: 11086.2
disk total throughput: 596.8 MiB/s
disk read iops: 6929.6
disk read throughput: 137.0 MiB/s
disk write iops: 4156.6
disk write throughput: 459.8 MiB/s
insert iops: 6883.0/s
insert throughput: 430.2 MiB/s
insert lat p50: 3us
insert lat p90: 5us
insert lat p99: 8us
insert lat p999: 23us
insert lat p9999: 153us
insert lat p99999: 235us
insert lat pmax: 345us
get iops: 17159.2/s
get miss: 5.16%
get throughput: 1.0 GiB/s
get hit lat p50: 4us
get hit lat p90: 367us
get hit lat p99: 859us
get hit lat p999: 2991us
get hit lat p9999: 6399us
get hit lat p99999: 6879us
get hit lat pmax: 6943us
get miss lat p50: 168us
get miss lat p90: 463us
get miss lat p99: 727us
get miss lat p999: 951us
get miss lat p9999: 1199us
get miss lat p99999: 1351us
get miss lat pmax: 1447us
```